backend/services/agent/requirements.txt
# ==== Core dependencies ====
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.8.0
psycopg2-binary==2.9.9
httpx==0.27.0
pgvector==0.2.5
sentence-transformers==3.0.1

# ==== Other common deps ====
redis==5.0.1
requests==2.31.0
typing_extensions>=4.8.0
Jinja2==3.1.2
gtts==2.5.1
python-dotenv==1.0.0
beautifulsoup4==4.12.3
lxml==5.2.2
pytesseract==0.3.10
pdfminer.six==20240706
python-magic==0.4.27
sentencepiece==0.2.0

# ==== Whisper (commented out for now) ====
# openai-whisper==20231117